# docs/swagger.yaml
basePath: /
definitions:
...
  ds.Request:
    properties:
      company_name:
        type: string
      consultation_place:
        type: string
      consultation_time:
        type: string
      endDate:
        type: string
      formationDate:
        type: string
      id:
        type: integer
      moderatorID:
        type: integer
      startDate:
        type: string
      status:
        type: string
      userID:
        type: integer
    type: object
  ds.StatusData:
    properties:
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Consultation app
  title: IT Services
  version: "1.0"
paths:
...
  /requests:
    get:
      description: Get all requests
      operationId: get-requests
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.Request'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Request'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Request'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Request'
      security:
      - ApiKeyAuth: []
      summary: Get Requests
      tags:
      - Requests
  /requests/{id}/moderator/update-status:
    put:
      consumes:
      - application/json
      description: Update request by moderator
      operationId: update-request-status-by-moderator
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: status info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/ds.StatusData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Request'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Request'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Request'
      security:
      - ApiKeyAuth: []
      summary: Update Request Status By Moderator
      tags:
      - Requests
  /requests/{id}/user/update-status:
    put:
      consumes:
      - application/json
      description: Update request status by user
      operationId: update-request-status-by-user
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Request'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Request'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Request'
      security:
      - ApiKeyAuth: []
      summary: Update Request Status By User
      tags:
      - Requests
  /requests/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete request by ID
      operationId: delete-request-by-id
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Request'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Request'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Request'
      security:
      - ApiKeyAuth: []
      summary: Delete Request by ID
      tags:
      - Requests
  /requests/update/{id}:
    put:
      consumes:
      - application/json
      description: Update request by ID
      operationId: update-request-by-id
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: request info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/ds.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Request'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Request'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Request'
      security:
      - ApiKeyAuth: []
      summary: Update Request by ID
      tags:
      - Requests
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"